+++ src/ft_string/ft_str/ft_atoi_base.c
@@ -6,7 +6,7 @@
 /*   By: bgoulard <bgoulard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/09 17:52:58 by bgoulard          #+#    #+#             */
-/*   Updated: 2024/05/21 17:36:58 by bgoulard         ###   ########.fr       */
+/*   Updated: 2024/06/13 16:54:39 by bgoulard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,10 +36,10 @@ static int	ft_is_valid(const char *base)
 
 int	ft_atoi_base(const char *str, const char *base)
 {
-	const int	base_len = ft_strlen(base);
-	int			i;
-	int			nb;
-	int			sign;
+	const size_t	base_len = ft_strlen(base);
+	int				i;
+	int				nb;
+	int				sign;
 
 	i = 0;
 	nb = 0;
